Windows specific kubelet flags in kubeadm-flags.env - Uses correct pause image for Windows - Omits systemd specific flags - Common build flags function to be used by Linux and Windows - Uses user configured image repository for Windows pause image
