+++ cmd/kubeadm/app/constants/constants.go
@@ -265,9 +265,6 @@ const (
 	// DefaultEtcdVersion indicates the default etcd version that kubeadm uses
 	DefaultEtcdVersion = "3.4.3-0"
 
-	// PauseVersion indicates the default pause image version for kubeadm
-	PauseVersion = "3.2"
-
 	// Etcd defines variable used internally when referring to etcd component
 	Etcd = "etcd"
 	// KubeAPIServer defines variable used internally when referring to kube-apiserver component
