+++ cmd/kubeadm/app/constants/constants_unix.go
@@ -21,4 +21,7 @@ package constants
 const (
 	// DefaultDockerCRISocket defines the default Docker CRI socket
 	DefaultDockerCRISocket = "/var/run/dockershim.sock"
+
+	// PauseVersion indicates the default pause image version for kubeadm
+	PauseVersion = "3.2"
 )
